Added Audio & Backgrounds
- Renamed files to be more coherent & to begin numbering with 0 instead of
  1
- Changed MOVE_SPEED to be an int
- Added Background enumeration
- Added background functionality to Graphics: loadBG(), which loads the
  background; closeBG(), which unloads the background(); manageBG(), which
  is called each frame to handle the background objects & parallax
  scrolling; bg variable, which stores the background image; bgObjects
  vector, which stores all background objects; bgPrefix, which is used in
  loading backgrounds
- Added needed #include's and forward declarations in Include.h
- Added Audio class: init() initializes sound effects and music; close()
  unloads sound effects and music; play() can be called to play a
  specified sound effect or music track; sfx and music are vectors which
  store the sfx and music objects; TOTAL_MUSIC and TOTAL_SFX store the
  total number of music tracks and sound effects; the remaining variables
  are used in file loading
- generateLevel() now calls functions to play level music and load level
  background

+++ Game.h
@@ -31,7 +31,7 @@ public:
 	static const float JUMP_MAX;
 	static const float JUMP_MULT;
 	static const float JUMP_START;
-	static const float MOVE_SPEED;
+	static const int MOVE_SPEED;
 	static const float ROTATION_SPEED;
 	static std::map<std::string, int> Controls;
 	static std::vector<int> entityOffset;
